Extracted common functions to utils.
